fix(processors): CI failures and second-round CodeRabbit findings
Registry attribute gap caught by CI's registry_pins test:
signaldb.signal and signaldb.processors.count were emitted but never
declared in otel/registry/signaldb.yaml, and the count was passed as a
usize instead of i64. Also KNOWN_ERROR_TYPES/throttle_of in the SDK
didn't know about the generated ProcessorError type.

Second round of CodeRabbit findings:
- ottl: Int()'s Double branch rejected non-finite and out-of-range
  values by saturating/zeroing instead of erroring.
- router: :test's per-processor statement results now carry the owning
  processor's name (every processor's statement indices restart at 0,
  so a multi-processor dry run previously returned ambiguous duplicate
  indices); OpenAPI/SDK/TS client regenerated for the new field.
- ui: the dry-run TestPanel could apply a slow, superseded response
  over the current input; requests now carry a revision guard.
- docs: documented the fail-closed first-load contract and the two
  execution-time 422 cases for :test.

+++ src/acceptor/src/handler/processors_apply.rs
@@ -79,7 +79,7 @@ macro_rules! apply_processors {
                 &tenant_context.dataset_id,
                 $signal,
             );
-            span.record("signaldb.processors.count", processors.len());
+            span.record("signaldb.processors.count", processors.len() as i64);
 
             async {
                 for processor in &processors {
